condominioaldia/requirements.txt
amqp==2.1.4
appdirs==1.4.3
billiard==3.5.0.2
celery==4.0.2
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==1.11
django-allauth==0.31.0
django-easy-pdf==0.1.1
django-filter==1.0.2
django-rest-auth==0.9.1
djangorestframework==3.6.2
enum34==1.1.6
functools32==3.2.3.post2
html5lib==1.0b10
httplib2==0.10.3
idna==2.6
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter-core==4.3.0
kombu==4.0.2
MySQL-python==1.2.5
nbformat==4.4.0
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.0
pyparsing==2.2.0
PyPDF2==1.26.0
python-dateutil==2.6.0
python-openid==2.2.5
pytz==2017.2
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
traitlets==4.3.2
urllib3==1.22
vine==1.1.3
webencodings==0.5.1
xhtml2pdf==0.2b1
